Move array rewriter functions to source file
Previously, most of the code for the theory of arrays rewriter was in
the header file, slowing down compilation. This commit moves the
functions to the body, replaces the ifdef guard in the header file with
a `pragma once` and reformats the code using clang-format. Note, this
commit removes an inline qualifier from the preRewrite function but
given the size of the function, this should not impact performance.