Treat tautologies as "sat" during unit propegation.
Previously, tautologies caused a failure if they were encountered during
unit propegation.

Important Fact: Tautologies cannot imply anything during unit
propegation, because they can never reduce to unit clauses modulo a
partial assignmnet.

In LRAT proofs the checker is instructed to look at clauses in a certain
order, and each of those clauses must be unit (or bottom) for the proof
to check.

In DRAT, the checker needs to look for the unit clauses on its own.

The natural way to combine functionality is to write a function which
checks whether some clause is unit. (e.g. returns true/false). We have
such a function, `check_unit_and_maybe_mark` (which also does some other
stuff, but we'll just consider the checking here).

The question then arises, what should the function do if the clause is
TAUT? From the LRAT perspective, it's fine if the function fails, but
from the DRAT perspective, failing would be bad, because the correct
behavior would be to just return "not a unit clause", and go check other
clauses.

We were failing, and now we're returning "not a unit"