Another DRAT/LRAT compatibility bug
There was a problem in which if a DRAT proof failed to meet the
specifified requirements in the final addition of the empty clause.
Specifically, the empty clause would not be an AT, and then resolution
search would trigger and fail because the empty clause can't resolve
with stuff. Now we just return false, causing operational DRAT checking
to start.